Add Canvas
Let's get something exciting going. First we are going to render a Canvas
from react-three-fiber. This sets up the HTML canvas element and will
wrap around all of our entire 3D scene.

So we'll replace the H1 with the Canvas. We don't see anything yet,
which makes sense. Let's add a bit of configuration real quick.
- colorManagement just makes the colors look nicer
- shadowMap let's us have nice shadows
- let's set the initial position of the camera