Add sphere
Now to make something show up. Let's just make a sphere to start. Every
object we make is going to follow the same basic pattern. First, we make
a mesh component. We can pass additional props to this like position,
scale, etc. For now we will stick with the defaults.

Then, as children to the mesh, we need a _geometry_ and a _material_.
We're gonna use a sphereBufferGeometry. The args we pass here correspond
to the args for Three.js's SphereBufferGeometry class (Show docs).
Radius, width segments, height segments, etc. We only care about these
first three. If we pass a small number of segments, we'll get something
real chunky. The higher we go, the smoother, but also the more intensive
it will be to render it. Let's try 32, which is kind of a lot but it'll
make it look nice. We can always lower it later if it is laggy.

Next is the material. We'll use a meshStandardMaterial, and set the
color to Colors.shape (this is defined in the utils file). Notice how it
looks black on the screen. That is because this is a lit material. Its
not a bug, the sphere is black because there is no light shining on it.