Add Lights
Let's fix that by adding some lights to the scene. First, add this
Lights component to the scene. Now, open up Lights.js so we can
implement it.

I already made this DirectionalLight component because I'm too lazy to
type all of this live.

So inside the Lights component, we'll start with a React Fragment. In
there, let's do an ambient light. This is like the base level of
lightness of the scene. We don't want it to be too bright, because we
want some shadows.

Next up, throw our Directional Light in there. You can think of this as
sun light. It has a direction it shines in, but the light rays basically
travel in a straight line.

This looks decent, but we're going to add a couple more. Adding multiple
light sources can create some depth to the scene.
Point lights are kinda like little lightbulbs in the scene. They have a
position and emit light in a sphere around that point. We'll do a
regular white one, and maybe a red one to mix it up.

The difference is subtle, but it'll be more noticeable later.