Add floor
So we have our shaded sphere, but it would be nice to have some sort of
floor for the sphere to cast a shadow on. Let's make a Floor component.
We need a mesh. We need to rotate it around the x axis 90 degrees.
Three.js uses radians instead of degrees, so instead of 90 here I am
putting Math.PI / 2. Don't worry if you don't know or remember
trigonometry. Just do what I do and copy it from stack overflow. We'll
also set the y position so its under the sphere. And set this
receiveShadow flag.

The planeBufferGeometry's args are width and height. We want it to be
pretty big.

The shadowMaterial is just a material that can have shadows cast on it.
Easy peasy. Now go ahead and add that to the scene... but we don't see
anything. What gives? Oh, we need to go back to our sphere and give it
the castShadow flag. Maybe let's do the receiveShadow flag too just in
case. BOOM. That is very exciting.