Fog and soft shadows
Two more small things.

One: Let's add some fog. We can really make this whatever color we want.
It'll give a slight tint to our scene, which adds depth.

Two: Call this softShadows function from drei. This is a helper to make
really nice, fuzzy shadows.

So that's it for the first section. We're going to take a 15 minute
break. You're going to go into breakout rooms where you can play around
with what we've got so far. Some ideas:
- Tweak colors and positions.
- Change the sphere to a different shape. Consult the three.js docs.
  torusBufferGeometry, boxBufferGeometry, etc.
- Change the amount and properties of the lights.

There should be one or two people in your room that can help out if you
have any questions or get stuck. I'll also post a link to a new codepen
with what I have so far if you want a fresh start or if you are joining
late.