Factor out Shape
For this section, we are going to make our scene more dynamic.

First let's cut out our shape from the scene, then go into Shape.js,
and paste it in.

Now in ShapeCluster.js, lets render a Shape inside of a group component.
A group is kind of like a genera purpose wrapper. We can apply
transforms to it like position, rotation, etc, and it will apply to
everything in the group.

Now, real quick, lets put ShapeCluster in our scene so we can see our
sphere again. Great, but lets get a few more spheres going.