Random cluster
Go into Shape, and give it a position prop.

Let's have ShapeCluster take a number prop, for how many shapes to
render. We can give it a default of 10. So we want to make 10 shapes,
and give each one a random position. A good way to calculate random
positions is with React's useMemo hook. This way we can get a random
position once for each sphere. when the component first mounts.

I am going to make a custom hook called usePositions. It will take the
number of positions we want to generate. We're going to call useMemo.
Inside here, we make an array as big as the number of positions we want.

Then, map the array. Each element should be an array containing [xyz]
coordinates. So in the map, we return an array. Now we just need to
generate some random values for x, y, and z.

There are a lot of different ways we could determine these. I'm going
to punch in some random ranges that look good, using this randomRange
helper function.

Now we have our array of positions, we can map each one to a Shape
element. Don't forget to give each one a key. Now we've got this weird
random blob of shapes. It isn't _super_ exciting yet, but we'll get
there. Let's animate the spheres to make them stretch and squish.