Random cluster
Go into Shape, and give it a position prop.

Let's have ShapeCluster take a number prop, for how many shapes to
render. We can give it a default of 10. So we want to make 10 shapes,
and give each one a random position. A good way to calculate random
positions is with React's useMemo hook. This way we can get a random
position once for each sphere. when the component first mounts.

Let's look at what we have here. Here is a function called
randomPosition. It returns an array of three random numbers, an xyz
coordinate.

I am going to make a custom hook called usePositions. It will take the
number of positions we want to generate. We're going to call useMemo.
Inside here, we are going to call this helper function "times". We pass
it a number, and a function, and it will call the function that many
times, returning an array of all the return values. So we just say,
times, randomPosition. That will give us this many random positions.

Now we have our array of positions, we can map each one to a Shape
element. Don't forget to give each one a key. Now we've got this weird
random blob of shapes. It isn't _super_ exciting yet, but we'll get
there. Let's animate the spheres to make them stretch and squish.