Basic Animation
Animation in react-three-fiber is interesting. I told you we'd be
writing declarative code, and that is mostly true. However, for stuff
like animation where we want to update every frame (16ms or so), we
actually want to circumvent React and handle it imperatively. Our scene
would slow to a crawl if we were rerendering a lot of React components
that often.

In the shape, we need a ref to the mesh. Then, we'll use a
react-three-fiber hook called useFrame. We can pass it a function that
will run every frame. The function will get passed a state object that
has a lot of stuff on it, but most importantly it has the time elapsed
since we started the scene.

Animation is just changing a property over time. If I just say
"meshref.current.position.y = time", we will see all of our shapes
slowly rise up. Nice, but we can do better.