Scale anim and metalness
In addition to up and down motion, we can also update the y-scale to
give it a stretchy effect.

While we are here, let's give our shapes a little metallic shine. With
the standard material we can specify some physical properties. You can
refer to the Three.js docs to see what they all are.