Animate group
So we've animated our shapes, but wouldn't it be cool if we could
animate the entire shape cluster? Let's try it out. Same as before, make
a ref, put it on the _group_, and then call the useFrame hook.

I'm thinking we'll want some back-and-forth motion again, so we'll use
Math.sin. Let's try rotating the entire group around the y axis.

Look at that! That's a pretty nice swivel we've got going on. Let's call
that the end of section two. Now, another break. During the break, if
you'd like, play around with the animations we made. Try adjusting some
of the coefficients, or using different easing functions. If you're
feeling adventurous, try animating different properties, like the scale
of the group, or rotate the spheres on the z-axis, etc.