OrbitControls
Now, I could stare at this for hours. But I would really love it if I
could interact with it in some way. Give me something to click on.

First, an easy one. Let's import OrbitControls from drei and toss that
in the bottom of our scene. Now we can zoom with the mouse wheel, rotate
with left click, and pan with right-click (or command click). Now our
scene really feels like a full 3D space we can explore.