First useSpring attempt
Then, we can import and use the useSpring hook from react-spring. This
is an amazing little animation library that works with the dom as well
as r3f (NOTE: Make sure the import is react-spring/three. This version
of the library is specific to r3f).

The object we pass to useSpring can have any properties we want on it.
We're going to give it a color, and make it change depending on the
active state. the hook returns a similar object, with special animated
properties. useSpring is smart enough to tween between lots of different
variable types, even these css colors.

We'll also pass it a config key. We imported these preset configs from
react-spring. I'm going to use config.wobbly because it looks fun, but
you can find others on the react-spring docs.

Now just use that color variable in the material.